+++ packages/invoice/src/Support/En16931/Casts/DeliveryPartyCast.php
@@ -7,7 +7,6 @@ namespace Moox\Invoice\Support\En16931\Casts;
 use Illuminate\Contracts\Database\Eloquent\CastsAttributes;
 use Illuminate\Database\Eloquent\Model;
 use Moox\Invoice\Support\En16931\Address;
-use Moox\Invoice\Support\En16931\DeliveryPartyStoredArray;
 use Moox\Invoice\Support\En16931\Party;
 
 /**
@@ -17,6 +16,9 @@ use Moox\Invoice\Support\En16931\Party;
  */
 class DeliveryPartyCast implements CastsAttributes
 {
+    /** @var list<string> */
+    private const FORBIDDEN_KEYS = ['vat_id', 'tax_number', 'contact'];
+
     public function get(Model $model, string $key, mixed $value, array $attributes): ?Party
     {
         if ($value === null) {
@@ -59,7 +61,15 @@ class DeliveryPartyCast implements CastsAttributes
      */
     public static function migrateStoredArray(?array $decoded): ?array
     {
-        return DeliveryPartyStoredArray::migrate($decoded);
+        if ($decoded === null) {
+            return null;
+        }
+
+        if (self::isLegacyAddressShape($decoded)) {
+            return self::wrapLegacyAddressShape($decoded);
+        }
+
+        return self::stripForbiddenPartyKeys($decoded);
     }
 
     /**
@@ -68,7 +78,20 @@ class DeliveryPartyCast implements CastsAttributes
      */
     public static function unwrapStoredArray(?array $decoded): ?array
     {
-        return DeliveryPartyStoredArray::unwrap($decoded);
+        if ($decoded === null) {
+            return null;
+        }
+
+        if (! isset($decoded['address']) || ! is_array($decoded['address'])) {
+            return $decoded;
+        }
+
+        $name = trim((string) ($decoded['name'] ?? ''));
+        if ($name !== '') {
+            return $decoded;
+        }
+
+        return $decoded['address'];
     }
 
     /**
@@ -76,7 +99,8 @@ class DeliveryPartyCast implements CastsAttributes
      */
     public static function isLegacyAddressShape(array $decoded): bool
     {
-        return DeliveryPartyStoredArray::isLegacyAddressShape($decoded);
+        return array_key_exists('country_code', $decoded)
+            && ! array_key_exists('address', $decoded);
     }
 
     /**
@@ -84,7 +108,9 @@ class DeliveryPartyCast implements CastsAttributes
      */
     private function partyFromStoredArray(array $decoded): ?Party
     {
-        $payload = DeliveryPartyStoredArray::partyPayloadFromStored($decoded);
+        $payload = self::isLegacyAddressShape($decoded)
+            ? self::wrapLegacyAddressShape($decoded)
+            : $decoded;
         $name = (string) ($payload['name'] ?? '');
         $addressPayload = $payload['address'] ?? null;
 
@@ -122,17 +148,68 @@ class DeliveryPartyCast implements CastsAttributes
             return null;
         }
 
-        return json_encode($this->persistedArray($party->name, $party->address), JSON_THROW_ON_ERROR);
+        return json_encode([
+            'name' => $party->name,
+            'address' => $party->address->toArray(),
+        ], JSON_THROW_ON_ERROR);
     }
 
     /**
+     * @param  array<string, mixed>  $decoded
      * @return array{name: string, address: array<string, mixed>}
      */
-    private function persistedArray(string $name, Address $address): array
+    private static function wrapLegacyAddressShape(array $decoded): array
     {
+        $name = (string) ($decoded['name'] ?? '');
+        $address = $decoded;
+        unset($address['name'], $address['vat_id'], $address['tax_number'], $address['contact']);
+
         return [
             'name' => $name,
-            'address' => $address->toArray(),
+            'address' => $address,
         ];
+    }
+
+    /**
+     * @param  array<string, mixed>  $decoded
+     * @return array<string, mixed>
+     */
+    private static function stripForbiddenPartyKeys(array $decoded): array
+    {
+        [$cleaned, $dirty] = self::unsetKeysIfPresent($decoded, self::FORBIDDEN_KEYS);
+
+        if (! isset($cleaned['address']) || ! is_array($cleaned['address'])) {
+            return $dirty ? $cleaned : $decoded;
+        }
+
+        [$address, $addressDirty] = self::unsetKeysIfPresent($cleaned['address'], self::FORBIDDEN_KEYS);
+        if (! $addressDirty) {
+            return $dirty ? $cleaned : $decoded;
+        }
+
+        $cleaned['address'] = $address;
+
+        return $cleaned;
+    }
+
+    /**
+     * @param  array<string, mixed>  $data
+     * @param  list<string>  $keys
+     * @return array{0: array<string, mixed>, 1: bool}
+     */
+    private static function unsetKeysIfPresent(array $data, array $keys): array
+    {
+        $dirty = false;
+
+        foreach ($keys as $key) {
+            if (! array_key_exists($key, $data)) {
+                continue;
+            }
+
+            unset($data[$key]);
+            $dirty = true;
+        }
+
+        return [$data, $dirty];
     }
 }
